deduplicate copy/paste open_pcm, centralize single mutex
