+++ src/alsa_backend/mod.rs
@@ -14,6 +14,7 @@
 // Mozilla Public License along with this program. If not, see
 // <https://www.gnu.org/licenses/> and <https://www.mozilla.org/MPL/2.0/>.
 
+mod alsa_pcm;
 #[cfg(not(feature = "threaded-alsa"))]
 pub mod buffermanager;
 #[cfg(not(feature = "threaded-alsa"))]
